分离各个面板
