+++ ln_drawer/src/widgets/panel.rs
@@ -1,3 +1,7 @@
+pub mod color_picker;
+pub mod debug_panel;
+pub mod side_docker;
+
 use ln_world::{Element, Handle, World};
 
 use crate::{
